refactor: move example tests out of test_workflow and into an example file
